### 2023-04-14 #5371 @editorialbot
[REVIEW]: PlantSimEngine: A Simulation Engine For The Soil-Plant-Atmosphere System
**Submitting author:** @VEZY (<a href="http://orcid.org/0000-0002-0808-1461">Rémi VEZY</a>)
**Repository:** https://github.com/VirtualPlantLab/PlantSimEngine.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.8.1
