### 2023-04-17 VEZY/PlantSimEngine.jl#32 @ashiklom
[REVIEW] Add complete (if basic) usage example to docs homepage and README
I suggest including 1-2 examples of complete usage on the README and documentation home page. Basically, what you've already done in `PlantBiophysics.jl`.

For another good example, see [`SpeedyWeather.jl`](https://github.com/SpeedyWeather/SpeedyWeather.jl#example-use).

https://github.com/openjournals/joss-reviews/issues/5371
