### 2023-04-17 VEZY/PlantSimEngine.jl#33 @ashiklom
[Review] Add "real" multi-component leaf/plant model example
The current multi-model example in Model-Switching (https://vezy.github.io/PlantSimEngine.jl/stable/model_switching/) is a bit esoteric. Unless you're good at software engineering already (which is not the intended audience of this package), that makes it a bit tricky to see the value of the approach.

I suggest including a more real example -- even if it's a really basic one.

https://github.com/openjournals/joss-reviews/issues/5371
