### 2023-04-17 VEZY/PlantSimEngine.jl#34 @ashiklom
[REVIEW] Justify (or tone-down) performance claims in paper and docs
The paper and documentation make multiple claims about performance without benchmarks or jusfitication. I am not convinced that simple models --- like the examples here --- are that much faster in their Julia implementations relative to good implementations in Python or R (especially because in practice, a lot of the performance-critical steps in R/Python are taking advantage of highly optimized Fortran/C routines). Some suggestions (don't have to do all of these; just some ideas):

(1) Cite existing Julia performance benchmarks to highlight why some kinds of operations in Julia are as fast or faster than interpreted (and sometimes, compiled!) languages. Besides the relevant engineering/CS literature on Julia, existing scientific programming efforts in Julia --- like the CliMA group --- may have good numbers for this.

(2) Re-frame some sections as, "Modularity / flexibility of Julia's design allows users to take advantage of performance-enhancing libraries like Floops.jl, DifferentialEquations.jl", etc.

(3) Include some performance statistics for simulations. These don't necessarily have to be comparisons --- most people don't care about something being the fastest, just that it's fast enough (especially when you add all the other great functionality that this package provides).

https://github.com/openjournals/joss-reviews/issues/5371
