### 2023-04-17 VEZY/PlantSimEngine.jl#35 @ashiklom
[REVIEW] Add CONTRIBUTING.md file
The README invites users to submit pull requests or contact the authors. However, a proper `CONTRIBUTING.md` file that follows community guidelines (e.g., [from GitHub](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)) would be better.

https://github.com/openjournals/joss-reviews/issues/5371
