### 2023-04-19 a05b28a
Implement traits for the inputs. Fix #36
