### 2023-04-20 a2bcbd0
Update run.jl
Make row iteration over meteo object more generic using the Tables interface
