### 2023-04-20 cff3b22
Update dimensions.jl
Same for the dimension checks
