### 2023-04-22 793e521
CompatHelper: bump compat for PlantMeteo to 0.4, (keep existing compat)
