+++ Project.toml
@@ -21,7 +21,7 @@ CSV = "0.10"
 DataFrames = "1"
 FLoops = "0.2"
 MultiScaleTreeGraph = "0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.10"
-PlantMeteo = "0.1.2, 0.2, 0.3"
+PlantMeteo = "0.1.2, 0.2, 0.3, 0.4"
 Tables = "1"
 Term = "1, 2"
 julia = "1.6"
