### 2023-04-24 83d5dcc
clean-up examples/dummy.jl
