### 2023-04-24 e52ddf4
Create ToyLAIModel.jl
