+++ examples/ToyLAIModel.jl
@@ -1,0 +1,34 @@
+
+# Declaring the process of LAI dynamic:
+@process "LAI_Dynamic" verbose = false
+
+
+# Declaring the model of LAI dynamic with its parameter values:
+struct ToyLAIModel <: AbstractLai_DynamicModel
+    max_lai::Float64
+    dd_incslope::Int
+    inc_slope::Float64
+    dd_decslope::Int
+    dec_slope::Float64
+end
+
+# Defining a method with keyword arguments and default values:
+ToyLAIModel(; max_lai=8.0, dd_incslope=500, inc_slope=70, dd_decslope=1000, dec_slope=20) = ToyLAIModel(max_lai, dd_incslope, inc_slope, dd_decslope, dec_slope)
+
+# Defining the inputs and outputs of the model:
+PlantSimEngine.inputs_(::ToyLAIModel) = (degree_days_cu=-999,)
+PlantSimEngine.outputs_(::ToyLAIModel) = (LAI=-Inf,)
+
+# Implementing the actual algorithm by adding a method to the run! function for our model:
+function PlantSimEngine.run!(::ToyLAIModel, models, status, meteo, constants=nothing, extra=nothing)
+    status.LAI =
+        models.LAI_Dynamic.max_lai *
+        (1 /
+         (1 + exp((models.LAI_Dynamic.dd_incslope - status.degree_days_cu) / models.LAI_Dynamic.inc_slope)) -
+         1 / (1 + exp((models.LAI_Dynamic.dd_decslope - status.degree_days_cu) / models.LAI_Dynamic.dec_slope))
+        )
+
+    if status.LAI < 0
+        status.LAI = 0
+    end
+end
