+++ test/test-simulation.jl
@@ -1,4 +1,4 @@
-# includet("../examples/dummy.jl")
+# include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))
 
 @testset "Check missing model" begin
     # No problem here:
