+++ docs/src/FAQ/translate_a_model.md
@@ -1,0 +1,158 @@
+# I want to use PlantSimEngine for my model
+
+```@setup mymodel
+using PlantSimEngine
+using CairoMakie
+using CSV, DataFrames
+include(joinpath(pkgdir(PlantSimEngine), "examples/ToyLAIModel.jl"))
+
+function lai_toymodel(degree_days_cu; max_lai=8.0, dd_incslope=500, inc_slope=70, dd_decslope=1000, dec_slope=20)
+    LAI = max_lai * (1 / (1 + exp((dd_incslope - degree_days_cu) / inc_slope)) - 1 / (1 + exp((dd_decslope - degree_days_cu) / dec_slope)))
+    if LAI < 0
+        LAI = 0
+    end
+    return LAI
+end
+
+meteo_day = CSV.read(joinpath(pkgdir(PlantSimEngine), "examples/meteo_day.csv"), DataFrame, header=18)
+# Note: meteo_day is defined below if you want to reproduce it, then use this to write it:
+# PlantMeteo.write_weather("examples/meteo_day.csv", meteo_day, duration = Dates.Day)
+```
+
+If you already have a model, you can easily use `PlantSimEngine` to couple it with other models with minor adjustments.
+
+## Toy LAI Model 
+
+### Model description
+
+Let's take an example with a simple LAI model that we define below:
+
+```julia
+"""
+Simulate leaf area index (LAI, m² m⁻²) for a crop based on the amount of degree-days since sowing with a simple double-logistic function.
+
+# Arguments
+
+- `degree_days_cu`: degree-days since sowing
+- `max_lai=8`: Maximum value for LAI
+- `dd_incslope=500`: degree-days at which we get the maximal increase in LAI
+- `inc_slope=5`: slope of the increasing part of the LAI curve
+- `dd_decslope=1000`: degree-days at which we get the maximal decrease in LAI
+- `dec_slope=2`: slope of the decreasing part of the LAI curve
+"""
+function lai_toymodel(degree_days_cu; max_lai=8.0, dd_incslope=500, inc_slope=70, dd_decslope=1000, dec_slope=20)
+    LAI = max_lai * (1 / (1 + exp((dd_incslope - degree_days_cu) / inc_slope)) - 1 / (1 + exp((dd_decslope - degree_days_cu) / dec_slope)))
+    if LAI < 0
+        LAI = 0
+    end
+    return LAI
+end
+```
+
+This model takes the number of days since sowing as input and returns the simulated LAI. We can plot the simulated LAI for a year:
+
+```@example mymodel
+using CairoMakie
+
+lines(1:1300, lai_toymodel.(1:1300), color=:green, axis=(ylabel="LAI (m² m⁻²)", xlabel="Days since sowing"))
+```
+
+### Changes for PlantSimEngine
+
+The model can be implemented using `PlantSimEngine` as follows:
+
+#### Define a process
+
+If the process of LAI dynamic is not implement yet, we can define it like so:
+
+```julia
+@process LAI_Dynamic
+```
+
+#### Define the model
+
+We have to define a structure for our model that will contain the parameters of the model:
+
+```julia
+struct ToyLAIModel <: AbstractLai_DynamicModel
+    max_lai::Float64
+    dd_incslope::Int
+    inc_slope::Float64
+    dd_decslope::Int
+    dec_slope::Float64
+end
+```
+
+We can also define default values for the parameters by defining a method with keyword arguments:
+
+```julia
+ToyLAIModel(; max_lai=8.0, dd_incslope=500, inc_slope=70, dd_decslope=1000, dec_slope=20) = ToyLAIModel(max_lai, dd_incslope, inc_slope, dd_decslope, dec_slope)
+```
+
[…]
