### 2023-04-24 ebd2bf8
up julia compat (v>1.7) because of last version of PlantMeteo
