+++ Project.toml
@@ -21,10 +21,10 @@ CSV = "0.10"
 DataFrames = "1"
 FLoops = "0.2"
 MultiScaleTreeGraph = "0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.10"
-PlantMeteo = "0.1.2, 0.2, 0.3, 0.4"
+PlantMeteo = "0.5"
 Tables = "1"
 Term = "1, 2"
-julia = "1.6"
+julia = "1.7"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
