+++ src/dependencies/hard_dependencies.jl
@@ -1,5 +1,5 @@
 function hard_dependencies(models; verbose::Bool=true)
-    dep_tree = Dict(
+    dep_graph = Dict(
         p => HardDependencyNode(
             i,
             p,
@@ -13,12 +13,12 @@ function hard_dependencies(models; verbose::Bool=true)
     for (process, i) in pairs(models) # for each model in the model list
         level_1_dep = dep(i) # we get the dependencies of the model
         length(level_1_dep) == 0 && continue # if there is no dependency we skip the iteration
-        dep_tree[process].dependency = level_1_dep
+        dep_graph[process].dependency = level_1_dep
         for (p, depend) in pairs(level_1_dep) # for each dependency of the model i
             if hasproperty(models, p)
                 if typeof(getfield(models, p)) <: depend
-                    parent_dep = dep_tree[process]
-                    push!(parent_dep.children, dep_tree[p])
+                    parent_dep = dep_graph[process]
+                    push!(parent_dep.children, dep_graph[p])
                     for child in parent_dep.children
                         child.parent = parent_dep
                     end
@@ -34,10 +34,10 @@ function hard_dependencies(models; verbose::Bool=true)
                     push!(dep_not_found, p => depend)
 
                     push!(
-                        dep_tree[process].missing_dependency,
+                        dep_graph[process].missing_dependency,
                         findfirst(x -> x == p, keys(level_1_dep))
                     ) # index of the missing dep
-                    # NB: we can retreive missing deps using dep_tree[process].dependency[dep_tree[process].missing_dependency]
+                    # NB: we can retreive missing deps using dep_graph[process].dependency[dep_graph[process].missing_dependency]
                 end
             else
                 if verbose
@@ -50,23 +50,23 @@ function hard_dependencies(models; verbose::Bool=true)
                 push!(dep_not_found, p => depend)
 
                 push!(
-                    dep_tree[process].missing_dependency,
+                    dep_graph[process].missing_dependency,
                     findfirst(x -> x == p, keys(level_1_dep))
                 ) # index of the missing dep
-                # NB: we can retreive missing deps using dep_tree[process].dependency[dep_tree[process].missing_dependency]
+                # NB: we can retreive missing deps using dep_graph[process].dependency[dep_graph[process].missing_dependency]
             end
         end
     end
 
-    roots = [AbstractTrees.getroot(i) for i in values(dep_tree)]
-    # Keeping only the trees with no common root nodes, i.e. remove trees that are part of a
-    # bigger dependency tree:
+    roots = [AbstractTrees.getroot(i) for i in values(dep_graph)]
+    # Keeping only the graphs with no common root nodes, i.e. remove graphs that are part of a
+    # bigger dependency graph:
     unique_roots = Dict{Symbol,HardDependencyNode}()
-    for (p, m) in dep_tree
+    for (p, m) in dep_graph
         if m in roots
             push!(unique_roots, p => m)
         end
     end
 
-    return DependencyTree(unique_roots, dep_not_found)
+    return DependencyGraph(unique_roots, dep_not_found)
 end
