### 2023-04-24 6d548e4
Update table_traits.jl
Add method for `DataFrameRow`
