+++ src/table_traits.jl
@@ -64,3 +64,4 @@ DataFormat(::Type{<:PlantMeteo.TimeStepRow}) = SingletonAlike()
 DataFormat(::Type{<:Nothing}) = SingletonAlike() # For meteo == Nothing
 DataFormat(T::Type{<:Any}) = error("Unknown data format: $T.\nPlease define a `DataFormat` method, e.g.: DataFormat(::Type{$T}) method.")
 DataFormat(x::T) where {T} = DataFormat(T)
+DataFormat(::Type{<:DataFrames.DataFrameRow}) = SingletonAlike()
