### 2023-04-24 6dd5f2d
Update examples + add tests
