+++ examples/ToyLAIModel.jl
@@ -13,10 +13,10 @@ struct ToyLAIModel <: AbstractLai_DynamicModel
 end
 
 # Defining a method with keyword arguments and default values:
-ToyLAIModel(; max_lai=8.0, dd_incslope=500, inc_slope=70, dd_decslope=1000, dec_slope=20) = ToyLAIModel(max_lai, dd_incslope, inc_slope, dd_decslope, dec_slope)
+ToyLAIModel(; max_lai=8.0, dd_incslope=800, inc_slope=110, dd_decslope=1500, dec_slope=20) = ToyLAIModel(max_lai, dd_incslope, inc_slope, dd_decslope, dec_slope)
 
 # Defining the inputs and outputs of the model:
-PlantSimEngine.inputs_(::ToyLAIModel) = (degree_days_cu=-999,)
+PlantSimEngine.inputs_(::ToyLAIModel) = (degree_days_cu=-Inf,)
 PlantSimEngine.outputs_(::ToyLAIModel) = (LAI=-Inf,)
 
 # Implementing the actual algorithm by adding a method to the run! function for our model:
