+++ README.md
@@ -60,7 +60,7 @@ include(joinpath(pkgdir(PlantSimEngine), "examples/ToyLAIModel.jl"))
 # Define the model:
 model = ModelList(
     ToyLAIModel(),
-    status=(degree_days_cu=1.0:2000.0,),
+    status=(degree_days_cu=1.0:2000.0,), # Pass the cumulated degree-days as input to the model
 )
 
 run!(model) # run the model
@@ -95,15 +95,14 @@ Of course you can plot the outputs quite easily:
 # ] add CairoMakie
 using CairoMakie
 
-f, ax, p = lines(model[:degree_days_cu], model[:LAI], color=:green, axis=(ylabel="LAI (m² m⁻²)", xlabel="Cumulated growing degree days since sowing (°C)"))
-save("examples/LAI_growth.png", f) # save the figure
+lines(model[:degree_days_cu], model[:LAI], color=:green, axis=(ylabel="LAI (m² m⁻²)", xlabel="Cumulated growing degree days since sowing (°C)"))
 ```
 
 ![LAI Growth](examples/LAI_growth.png)
 
 ### Model coupling
 
-Model coupling is done automatically by the package, and is based on the dependency graph between the models. To couple model, we just have to add them to the `ModelList`. For example, let's couple the `ToyLAIModel` with a simple Beer's law model for light interception, and a simple model that computes the accumulation of degree days:
+Model coupling is done automatically by the package, and is based on the dependency graph between the models. To couple models, we just have to add them to the `ModelList`. For example, let's couple the `ToyLAIModel` with a model for light interception based on Beer's law:
 
 ```julia
 # ] add PlantSimEngine, DataFrames, CSV
@@ -112,18 +111,38 @@ using PlantSimEngine, PlantMeteo, DataFrames, CSV
 # Include the model definition from the examples folder:
 include(joinpath(pkgdir(PlantSimEngine), "examples/ToyLAIModel.jl"))
 include(joinpath(pkgdir(PlantSimEngine), "examples/light.jl"))
-include(joinpath(pkgdir(PlantSimEngine), "examples/ToyDegreeDays.jl"))
 
 # Import the example meteorological data:
 meteo_day = CSV.read(joinpath(pkgdir(PlantSimEngine), "examples/meteo_day.csv"), DataFrame, header=18)
 
-# Define the model:
+# Define the list of models for coupling:
 model = ModelList(
     ToyLAIModel(),
     Beer(0.6),
-    status=(degree_days_cu=cumsum(meteo_day[:, :degree_days]),),
+    status=(degree_days_cu=cumsum(meteo_day[:, :degree_days]),),  # Pass the cumulated degree-days as input to `ToyLAIModel`, this could also be done using another model
 )
+```
 
+The `ModelList` couples the models by automatically computing the dependency graph of the models. The resulting dependency graph is:
+
+```
+╭──── Dependency graph ──────────────────────────────────────────╮
+│  ╭──── LAI_Dynamic ─────────────────────────────────────────╮  │
+│  │  ╭──── Main model ────────╮                              │  │
+│  │  │  Process: LAI_Dynamic  │                              │  │
+│  │  │  Model: ToyLAIModel    │                              │  │
+│  │  │  Dep: nothing          │                              │  │
+│  │  ╰────────────────────────╯                              │  │
+│  │                  │  ╭──── Soft-coupled model ─────────╮  │  │
[…]
