### 2023-04-25 8c631dd
Add examples to docs landing page regarding #32
