### 2023-04-25 839a19c
Fix computation for fitting Beer + update tests
