### 2023-04-25 2adeb2c
Update runtests.jl
test doc only on v1.8 because printing of NamedTuples changed in v1.9
