+++ test/runtests.jl
@@ -46,7 +46,7 @@ include(joinpath(pkgdir(PlantSimEngine), "examples/Beer.jl"))
         include("test-fitting.jl")
     end
 
-    if VERSION >= v"1.8"
+    if VERSION == v"1.8"
         # Error formating changed in Julia 1.8 (or was it 1.7?), so the doctest
         # that returns an error in PlantSimEngine.check_dimensions(models, w)
         # fails in Julia 1.6. So we test the doctests only in Julia 1.8 and later.
