### 2023-04-25 VEZY/PlantSimEngine.jl#32 @VEZY
Hey @ashiklom, thanks for the review. I added two basic examples to the [README](https://github.com/VEZY/PlantSimEngine.jl/blob/main/README.md) + the [landing page](https://vezy.github.io/PlantSimEngine.jl/dev/) of the doc. Is this what you expected ? If not, let me know I'll be happy to improve it further.

PS: the docs in [SpeedyWeather.jl](https://speedyweather.github.io/SpeedyWeather.jl/dev/) are down (very nice package btw, I would love to use it at some point!).
