### 2023-04-25 9e88474
Add tests for Toy models
