### 2023-04-25 88abf14
Implement traits for parallel execution over time-steps or objects (see #21)
