+++ docs/src/fitting.md
@@ -7,7 +7,7 @@ meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0)
 m = ModelList(Beer(0.6), status=(LAI=2.0,))
 run!(m, meteo)
 
-df = DataFrame(PPFD=m[:PPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
+df = DataFrame(aPPFD=m[:aPPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
 ```
 
 ## The fit method
@@ -61,7 +61,7 @@ run!(m, meteo)
 Now we can define the "data" to fit the model using the simulated `PPFD` values:
 
 ```@example usepkg
-df = DataFrame(PPFD=m[:PPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
+df = DataFrame(aPPFD=m[:aPPFD][1], LAI=m.status.LAI[1], Ri_PAR_f=meteo.Ri_PAR_f[1])
 ```
 
 And finally we can fit the model using the `fit` method:

[…]
