### 2023-04-26 b89b43e
Update process_generation.jl
Add doc about how to define a process without the macro + add info about traits
