### 2023-04-26 f58d2b2
Update inputs.md
Add compat info for PlantMeteo.row_from_parent
