### 2023-04-26 bef05f9
Update implement_a_model.md
Add info about parallel traits
