### 2023-04-26 98d38bd
Update implement_a_process.md
Growth model is now using `ToyAssimGrowthModel`
