+++ docs/src/extending/implement_a_process.md
@@ -17,7 +17,7 @@ To implement a new process, we need to define an abstract structure that will he
 For example, the photosynthesis process in [PlantBiophysics.jl](https://github.com/VEZY/PlantBiophysics.jl) is declared using just this tiny line of code:
 
 ```julia
-@process photosynthesis
+@process "photosynthesis"
 ```
 
 If we want to simulate the growth of a plant, we could add a new process called `growth`:
@@ -28,96 +28,110 @@ If we want to simulate the growth of a plant, we could add a new process called 
 
 And that's it! Note that the function guides you in the steps you can make after creating a process. Let's break it up here.
 
-So what you just did is to create a new process called `growth`. By doing so, you created a new abstract structure called `AbstractGrowthModel`. It is used as a supertype for the types used for model implementation. This abstract type is always named using the process name in title case (using `titlecase()`), prefixed with `Abstract` and suffixed with `Model`.
+!!! note
+    If you know what you're doing, you can directly define a process by hand just by defining an abstract type that is a subtype of `AbstractModel`:
+    ```julia
+    abstract type AbstractGrowthModel <: PlantSimEngine.AbstractModel end
+    ```
+    And by adding a method for the `process_` function that returns the name of the process:
+    ```julia
+    PlantSimEngine.process_(::Type{AbstractGrowthModel}) = :growth
+    ```
+    But this way, you don't get the nice tutorial adapted to your process 😃.
+
+So what you just did is to create a new process called `growth`. By doing so, you created a new abstract structure called `AbstractGrowthModel`, which is used as a supertype of the models. This abstract type is always named using the process name in title case (using `titlecase()`), prefixed with `Abstract` and suffixed with `Model`.
 
 !!! note
     If you don't understand what a supertype is, no worries, you'll understand by seeing the examples below
 
 ## Implement a new model for the process
 
-To better understand how models are implemented, you can read the detailed instructions from the [previous section](@ref model_implementation_page). But for the sake of completeness, we'll implement a growth model here.
+To better understand how models are implemented, you can read the detailed instructions from the [next section](@ref model_implementation_page). But for the sake of completeness, we'll implement a growth model here.
 
-This growth model needs the carbohydrate assimilation that we could compute using *e.g.* the coupled energy balance process from `PlantBiophysics.jl`. Then the model removes the maintenance respiration and the growth respiration from that source of carbon, and increments the leaf biomass by the remaining carbon offer.
+This growth model needs the carbohydrate assimilation that we could compute using *e.g.* the coupled energy balance process from `PlantBiophysics.jl`. Then the model removes the maintenance respiration and the growth respiration from that source of carbon, and increments the plant biomass by the remaining carbon offer.
 
-Let's implement this model below:
+The model is available in the example script [ToyAssimGrowthModel.jl](https://github.com/VEZY/PlantSimEngine.jl/blob/main/examples/ToyAssimGrowthModel.jl), and is reproduced below:
 
 ```@example usepkg
[…]
