+++ docs/src/model_switching.md
@@ -1,40 +1,28 @@
 # Model switching
 
 ```@setup usepkg
-using PlantSimEngine, PlantMeteo
-include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))
-meteo = Atmosphere(T = 20.0, Wind = 1.0, P = 101.3, Rh = 0.65)
-m = ModelList(
-    Process1Model(2.0), 
-    Process2Model(),
-    Process3Model(),
-    Process4Model(),
-    Process5Model(),
-    Process6Model(),
-    Process7Model(),
-    status = (var0=1.0,)
+using PlantSimEngine, PlantMeteo, CSV, DataFrames
+include(joinpath(pkgdir(PlantSimEngine), "examples/ToyLAIModel.jl"))
+include(joinpath(pkgdir(PlantSimEngine), "examples/Beer.jl"))
+include(joinpath(pkgdir(PlantSimEngine), "examples/ToyAssimGrowthModel.jl"))
+include(joinpath(pkgdir(PlantSimEngine), "examples/ToyRUEGrowthModel.jl"))
+
+meteo_day = CSV.read(joinpath(pkgdir(PlantSimEngine), "examples/meteo_day.csv"), DataFrame, header=18)
+ 
+models = ModelList(
+    ToyLAIModel(),
+    Beer(0.5),
+    ToyRUEGrowthModel(0.2),
+    status=(degree_days_cu=cumsum(meteo_day.degree_days),),
 )
-run!(m, meteo)
-struct AnotherProcess1Model <: AbstractProcess1Model
-    a
-    b # this is a new parameter
-end
-PlantSimEngine.inputs_(::AnotherProcess1Model) = (var1=-Inf, var2=-Inf, var10=-Inf,)
-PlantSimEngine.outputs_(::AnotherProcess1Model) = (var3=-Inf,)
-function PlantSimEngine.run!(::AnotherProcess1Model, models, status, meteo, constants=nothing, extra=nothing)
-    status.var3 = models.process1.a + status.var1 * status.var2 + status.var10
-end
-m2 = ModelList(
-    AnotherProcess1Model(2.0, 0.5), 
-    Process2Model(),
-    Process3Model(),
-    Process4Model(),
-    Process5Model(),
-    Process6Model(),
-    Process7Model(),
-    status = (var0=1.0, var10=2.0)
+run!(models, meteo_day)
+models2 = ModelList(
+    ToyLAIModel(),
+    Beer(0.5),
+    ToyAssimGrowthModel(),
+    status=(degree_days_cu=cumsum(meteo_day.degree_days),),
 )
-run!(m2, meteo)
+run!(models2, meteo_day)
 ```
 
 One of the main objective of PlantSimEngine is allowing users to switch between model implementations for a given process **without making any change to the code**. 
@@ -45,25 +33,26 @@ The package was carefully designed around this idea to make it easy and computat
 
 The `ModelList` is a container that holds a list of models, their parameter values, and the status of the variables associated to them.
 
-Model coupling is done by adding models to the `ModelList`. Let's create a `ModelList` with the seven models from the example script `examples/dummy.jl`:
+Model coupling is done by adding models to the `ModelList`. Let's create a `ModelList` with several models from the example scripts in the [`examples`](https://github.com/VEZY/PlantSimEngine.jl/blob/master/examples/) folder:
+
+Importing the models from the scripts:
 
 ```julia
 using PlantSimEngine
-include(joinpath(pkgdir(PlantSimEngine), "examples/dummy.jl"))
+include(joinpath(pkgdir(PlantSimEngine), "examples/ToyLAIModel.jl"))
[…]
