### 2023-04-27 #5371 @crvernon
🚨 MIDWEEK RALLY 🚨

It looks like @ashiklom made great progress on their pass of the review!  Thank you!  Lots of open issues for @VEZY to address and discuss!

@tpoisot and @tomyun - Could you please provide an update to how things are going here? As always, let me know if you have any questions!

Have a great day!
