### 2023-04-27 #5371 @VEZY
Yes thank you @ashiklom ! Much improvements coming to the package thanks to your review ! @crvernon, just a question about the issues: when I feel that I completed the reviewer's request, should I close the issue or let it open for further discussion ? For the moment, I let it open and leave a comment (see: https://github.com/VEZY/PlantSimEngine.jl/issues/32), but I feel that it may be confusing.
