### 2023-04-27 294480f
Update model_switching.md
Improve writing and add info about parallel warning #33
