### 2023-04-27 VEZY/PlantSimEngine.jl#34 @VEZY
@ashiklom, can you point me to any publication demonstrating performance from the CliMA group ? I found [this page](https://clima.caltech.edu/publications/)  but there are so many publications there...

We show a comparison with interpreted languages in the paper we are preparing for PlantBiophysics, but it does not represent the entire field of plant modelling, so it's really relevant here. So I'm preparing a little benchmark with the new toy models I added in the examples fodler.
