### 2023-04-27 VEZY/PlantSimEngine.jl#34 @ashiklom
I think just the simple comparison from the PlantBiophysics paper that you describe would be sufficient. Even better than anything from CLIMA, I think, because you are showing a direct comparison for exactly the computations that this package is doing.

If you will not have those comparisons for some time, it's also fine to just tweak the language to say something like, "Julia can achieve significantly better performance than typical interpreted languages" and cite _any_ studies or websites that have shown this. The performance is not the main selling point of this package, so this doesn't have to be perfect.
