+++ docs/paper/paper.bib
@@ -1,10 +1,30 @@
 
+@article{duursma_plantecophys_2015,
+	title = {Plantecophys - {An} {R} {Package} for {Analysing} and {Modelling} {Leaf} {Gas} {Exchange} {Data}},
+	volume = {10},
+	issn = {1932-6203},
+	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0143346},
+	doi = {10.1371/journal.pone.0143346},
+	abstract = {Here I present the R package 'plantecophys', a toolkit to analyse and model leaf gas exchange data. Measurements of leaf photosynthesis and transpiration are routinely collected with portable gas exchange instruments, and analysed with a few key models. These models include the Farquhar-von Caemmerer-Berry (FvCB) model of leaf photosynthesis, the Ball-Berry models of stomatal conductance, and the coupled leaf gas exchange model which combines the supply and demand functions for CO2 in the leaf. The 'plantecophys' R package includes functions for fitting these models to measurements, as well as simulating from the fitted models to aid in interpreting experimental data. Here I describe the functionality and implementation of the new package, and give some examples of its use. I briefly describe functions for fitting the FvCB model of photosynthesis to measurements of photosynthesis-CO2 response curves ('A-Ci curves'), fitting Ball-Berry type models, modelling C3 photosynthesis with the coupled photosynthesis-stomatal conductance model, modelling C4 photosynthesis, numerical solution of optimal stomatal behaviour, and energy balance calculations using the Penman-Monteith equation. This open-source package makes technically challenging calculations easily accessible for many users and is freely available on CRAN.},
+	language = {en},
+	number = {11},
+	urldate = {2020-02-23},
+	journal = {PLOS ONE},
+	author = {Duursma, Remko A.},
+	month = nov,
+	year = {2015},
+	note = {Number: 11},
+	keywords = {Photosynthesis, Carbon dioxide, C3 photosynthesis, C4 photosynthesis, Curve fitting, Leaves, Stomata, Supply and demand},
+	pages = {e0143346},
+	file = {Duursma_2015_Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data.pdf:/Users/vezy/Zotero/storage/I7D8QT7U/Duursma_2015_Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data.pdf:application/pdf;Snapshot:/Users/vezy/Zotero/storage/EHXCTFFA/article.html:text/html},
+}
+
 @article{brisson_stics_1998,
 	title = {{STICS}: a generic model for the simulation of crops and their water and nitrogen balances. {I}. {Theory} and parameterization applied to wheat and corn},
 	volume = {18},
 	shorttitle = {{STICS}},
 	url = {https://hal.inrae.fr/hal-02698682},
-	doi = {10.1051/agro:19980501},
+	doi = {https://doi.org/10.1051/agro:19980501},
 	number = {5-6},
 	urldate = {2021-06-20},
 	journal = {Agronomie},
@@ -21,7 +41,6 @@ Publisher: EDP Sciences},
 	title = {{OpenAlea}: a visual programming and component-based software platform for plant modelling},
 	volume = {35},
 	url = {https://doi.org/10.1071/FP08084},
-	doi = {10.1071/fp08084},
[…]
