### 2023-04-27 VEZY/PlantSimEngine.jl#34 @VEZY
Thank you for the quick answer ! I propose the changes in acdfc72361d6fbbdd72f6fba678af9f775c32093.
What I did:
- add a benchmark script so people can verify by themselves quickly, and put the results in the paper
- add link to benchmark of PlantBiophysics vs plantecophys (R implementation)
- toned-down some words that you identified: hassle-free, effortlessy...
- also not related to this issue but I added a citation to the paper Roesch et al. (2023) that nicely outlines the issues that we have in biology and how Julia can be a great tool to tackle them
