### 2023-04-27 ff8b01a
Update parallel_traits.jl
Models are object dependents by default !
