### 2023-04-27 3e89346
Update model_execution.md
replace runnable by executable
