### 2023-05-02 371b81a
Update paper.md
Minor suggestions related to the "Software paper" section from review of @ashiklom:

- Clarify that PlantSimEngine does not itself define any specific plant models
- fix inconsistent use of capitalization in Summary bullets
