+++ docs/paper/paper.md
@@ -53,9 +53,13 @@ PlantSimEngine is a state-of-the-art plant simulation platform that offers signi
 
 The use of Julia programming language in PlantSimEngine allows for quick and easy prototyping of models compared to compiled languages (STICS, APSIM, DSSAT) and can achieve significantly better performance than typical interpreted languages (*e.g.* Python used in OpenAlea), without the need for translation into another compiled language. For example PlantBiophysics.jl, a Julia package that implements ecophysiological models using PlantSimEngine.jl is [38649 times faster](https://vezy.github.io/PlantBiophysics-paper/notebooks_performance_Fig5_PlantBiophysics_performance/) than the same implementation in R in plantecophys [@duursma_plantecophys_2015].
 
-Most of the nice features of the package are provided by the Julia programming language, which is a high-level, high-performance, dynamic programming language used in many scientific fields and particularly well fitted for modelling in biology [@roesch_julia_2023].
+The package's impressive features are primarily provided by the Julia programming language. Julia is a high-level, high-performance, and dynamic programming language widely used in various scientific fields, and particularly for biology [@roesch_julia_2023]. The Julia community has recently developed remarkable tools such as Cropbox.jl [@yun_cropbox_2022], and CliMA Land [@wang_testing_2021].
 
-The automatic computation of the dependency graph of models and the uniform API enables users to switch models without having to modify any underlying code. This is a significant advantage over existing tools for the development of complex soil-plant-atmosphere models. For example OpenAlea provides a way to build a dependency tree for soft-dependencies, *i.e.* independent models that are coupled via inputs and outputs, but not for hard-dependencies, *i.e.* models that make an explicit, hard-coded call to another model. PlantSimEngine leverages Julia's multiple-dispatch to automatically compute the full dependency graph at compile time, including hard-dependencies. PlantSimEngine also offers a simple and generic API to define methods for model calibration that are used to programmatically calibrate models and sub-models.
+Cropbox.jl is a declarative crop modeling framework. It shares the same objective with PlantSimEngine, which is to allow users to define models without having to manage the complexity of the underlying simulations. One additional feature of PlantSimEngine is that it is not limited to the crop scale. It can be used for land surface models (LSM) down to functional-structural plant models (FSPM) thanks to its compatibility with MultiScaleTreeGraph.jl. PlantSimEngine is also faster because it is developed purely in Julia, whereas Cropbox.jl uses the Python package SymPy to run the simulations.
+
+CliMA Land is a next generation LSM that can simulate the Soil-Plant-Atmosphere continuum on a large scale. One interesting feature from PlantSimEngine that could be useful for CliMA Land is the ease of model coupling, including the concepts of hard and soft-coupled models, and automatic computation of the dependency graph. 
+
+The automatic computation of the dependency graph of models and the uniform API enables users to switch models without having to modify any underlying code. This is a significant advantage over existing tools for the development of complex soil-plant-atmosphere models. For example, OpenAlea provides a way to build a dependency tree for soft-dependencies, *i.e.* independent models that are coupled via inputs and outputs, but not for hard-dependencies, *i.e.* models that make an explicit, hard-coded call to another model. PlantSimEngine leverages Julia's multiple-dispatch to automatically compute the full dependency graph at compile time, including hard-dependencies. PlantSimEngine also offers a simple and generic API to define methods for model calibration that are used to programmatically calibrate models and sub-models.
 
 Additionally, PlantSimEngine offers the ability to reduce the number of degrees of freedom by fixing variables, passing measurements, or using a simpler model for a given process, making it an easier and more flexible solution for plant simulation.
 
