+++ docs/paper/paper.bib
@@ -139,3 +139,30 @@ Publisher: EDP Sciences},
 	year = {2020},
 	note = {arXiv: 2001.07625},
 }
+
+@article{yun_cropbox_2022,
+	title = {Cropbox: a declarative crop modelling framework},
+	volume = {5},
+	issn = {2517-5025},
+	url = {https://doi.org/10.1093/insilicoplants/diac021},
+	doi = {10.1093/insilicoplants/diac021},
+	abstract = {We introduce Cropbox, a novel modelling framework that supports various aspects of crop modelling in a unique yet concise style. Building a crop model can be easily riddled with technical details looking trivial at first but later becoming major obstacles that hamper the whole development or application process. This is particularly the case when implementing models from scratch without relying on an established framework. Cropbox adopts a declarative approach providing a domain-specific language to reduce technical debt and assist modellers to focus on high-level abstraction formed by relations between variables and enclosing systems, rather than tinkering with low-level implementation details. The syntax of Cropbox framework is based on the Julia programming language and is deliberately constrained to avoid unintended side effects caused by common mistakes while its architecture remains open to extension. We highlight key capabilities of the framework through case studies featuring a leaf gas-exchange model and a whole-plant simulation model. We also illustrate potential extensions for supporting functional-structural plant modelling by demonstrating a 3D root architectural model as an example.},
+	number = {1},
+	journal = {in silico Plants},
+	author = {Yun, Kyungdahm and Kim, Soo-Hyung},
+	month = dec,
+	year = {2022},
+	note = {\_eprint: https://academic.oup.com/insilicoplants/article-pdf/5/1/diac021/49420432/diac021.pdf},
+}
+
+@article{wang_testing_2021,
+	title = {Testing stomatal models at the stand level in deciduous angiosperm and evergreen gymnosperm forests using {CliMA} {Land} (v0.1)},
+	volume = {14},
+	url = {https://gmd.copernicus.org/articles/14/6741/2021/},
+	doi = {10.5194/gmd-14-6741-2021},
+	number = {11},
+	journal = {Geoscientific Model Development},
+	author = {Wang, Y. and Köhler, P. and He, L. and Doughty, R. and Braghiere, R. K. and Wood, J. D. and Frankenberg, C.},
+	year = {2021},
+	pages = {6741--6763},
+}
