### 2023-05-03 b0b314b
Update index and readme with corrections provided in paper
