### 2023-05-03 #5371 @VEZY
@ashiklom, I believe I answered all your suggestions / requests:

- I closed the issues you proposed:
    - Adding a CONTRIBUTING.md file
    - Adding better examples in the package documentation
    - Adding benchmarks for performance
- I fixed issues in PlantBiophysics (see v0.10.0)
- Tone down hyperbolic language
- Fix issues in capitalisation for bullet points + other minor fixes
- Add reference to CliMA Land (I also added a ref to Cropbox.jl)

Thank you for your suggestions, it significantly improved the package!
Let me know if you have any more requests.
