### 2023-05-17 e700e39
Add nsteps (closes #43)
