+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PlantSimEngine"
 uuid = "9a576370-710b-4269-adf9-4f603a9c6423"
 authors = ["Rémi Vezy <VEZY@users.noreply.github.com> and contributors"]
-version = "0.7.0"
+version = "0.8.0"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
