### 2023-05-19 #5371 @VEZY
@tomyun Hoo OK, sorry I misunderstood how you use SymPy in Cropbox.jl.

Would it be safe to say this for the difference between the two? :

- Cropbox rather defines the system as a set of continous models, whereas PlantSimEngine defines it as a discrete system
- Cropbox implements a DSL for crops with a declarative modelling framework, whereas PlantSimEngine uses a more functional approach
- Both can be used for FSP models implementations (I am not sure to have well understood how structure is handled in Cropbox). PlantSimEngine is compatible with the MultiScaleTreeGraph for multiple scales computations (e.g. one computation happening at the plant scale, another one at the leaf scale, for each leaf).

Also, I'm not sure to understand why LeagGasExchange is way slower than PlantBiophysics then. Maybe we are doing something wrong with our benchmark [here](https://vezy.github.io/PlantBiophysics-paper/notebooks_performance_Fig5_PlantBiophysics_performance/) ? Would you mind to have a look ? Maybe is it coming from the fact that we have type stability all the way in PlantSimEngine (even for the weather) ? Or
is it coming from the dependency graph building step ? We build it when calling `run!`, so at each iteration `N` in the benchmark, as for Cropbox I guess?
