### 2023-06-04 #5371 @tomyun
>- Cropbox rather defines the system as a set of continous models, whereas PlantSimEngine defines it as a discrete system
>- Cropbox implements a DSL for crops with a declarative modelling framework, whereas PlantSimEngine uses a more functional approach

Cropbox also simulates the model in a discrete time step. I believe the main difference comes from its DSL approach that forces an explicit declaration of variables. A motivation behind this approach is to prevent developers from writing a large opaque function that is often required for describing update logic in many other modeling frameworks.

>- Both can be used for FSP models implementations (I am not sure to have well understood how structure is handled in Cropbox). PlantSimEngine is compatible with the MultiScaleTreeGraph for multiple scales computations (e.g. one computation happening at the plant scale, another one at the leaf scale, for each leaf).

Cropbox was initially designed for building conventional crop models with a rather static structure, but it's been extended to support dynamic generation of internal structure to implement dynamic structures, e.g., individual leaf development/growth and 3D root structure consists of small root segments. Yet, it's not fully fledged as other dedicated FSPM frameworks and we'd definitely like to work more on this area.

> Also, I'm not sure to understand why LeagGasExchange is way slower than PlantBiophysics then. Maybe we are doing something wrong with our benchmark [here](https://vezy.github.io/PlantBiophysics-paper/notebooks_performance_Fig5_PlantBiophysics_performance/) ? Would you mind to have a look ? Maybe is it coming from the fact that we have type stability all the way in PlantSimEngine (even for the weather) ? Or
is it coming from the dependency graph building step ? We build it when calling run!, so at each iteration N in the benchmark, as for Cropbox I guess?

There should be many reasons behind its relatively slow performance when compared to vanilla Julia implementations. Type stability is one of major bottlenecks as we wanted to use Cropbox in an interactive environment where we can't reliably expect data types, e.g., redefining model definition, extracting arbitrary variable outputs, etc.

With that, I believe there is a clear difference between models included in the benchmark how they implement gas-exchange models. For instance, LeafGasExchange.jl tries to solve Ci and Tl more aggressively using a bisection method which is quite reliable but slow. I believe PhotosynEB in plantecophys doesn't optimize these variables after an intial pass. On the other hand, PlantBiophysics.jl seems to solve Ci in a unique way relying on two analytics solutions of Ci for separte limiting pathways whereas LeafGasExchange.jl tries to solve it numerically without assuming a specific strucute of limiting pathways.
