+++ docs/paper/paper.md
@@ -55,7 +55,9 @@ The use of Julia programming language in PlantSimEngine allows for quick and eas
 
 The package's impressive features are primarily provided by the Julia programming language. Julia is a high-level, high-performance, and dynamic programming language widely used in various scientific fields, and particularly for biology [@roesch_julia_2023]. The Julia community has recently developed remarkable tools such as Cropbox.jl [@yun_cropbox_2022], and CliMA Land [@wang_testing_2021].
 
-Cropbox.jl is a declarative crop modeling framework. It shares the same objective with PlantSimEngine, which is to allow users to define models without having to manage the complexity of the underlying simulations. One additional feature of PlantSimEngine is that it is not limited to the crop scale. It can be used for land surface models (LSM) down to functional-structural plant models (FSPM) thanks to its compatibility with MultiScaleTreeGraph.jl. PlantSimEngine is also faster because it is developed purely in Julia, whereas Cropbox.jl uses the Python package SymPy to run the simulations.
+Cropbox.jl is a declarative crop modeling framework that share a common objective with PlantSimEngine: simplifying the process of defining models by eliminating the complexities of underlying simulations. While both frameworks can be applied across a wide range of scales, including land surface models (LSM), crop models, and functional-structural plant models (FSPM), they differ in their implementation strategies. 
+
+Cropbox.jl defines a domain-specific language (DSL) tailored specifically for crop modeling, providing an intuitive approach for model definition. On the other hand, PlantSimEngine prioritizes speed as a fundamental aspect, ensuring efficient computations and enabling scalability right from the start. This is accomplished through meticulous considerations, such as type stability and pre-allocation to optimize performance. Type stability in PlantSimEngine ensures that variables maintain consistent types throughout computations, which leads to efficient execution. Additionally, the compatibility of PlantSimEngine with MultiScaleTreeGraph.jl [@vezy_multiscaletreegraphjl_2023] further enhances its scalability and facilitates multi-scale computations. MultiScaleTreeGraph.jl provides a powerful framework for representing and analyzing hierarchical structures, enabling PlantSimEngine to seamlessly handle complex systems spanning multiple scales.
 
 CliMA Land is a next generation LSM that can simulate the Soil-Plant-Atmosphere continuum on a large scale. One interesting feature from PlantSimEngine that could be useful for CliMA Land is the ease of model coupling, including the concepts of hard and soft-coupled models, and automatic computation of the dependency graph. 
 
