### 2023-06-07 #5371 @VEZY
@tomyun thank you for your explanations, we'll definitely have to talk some time ! I'll update how we explain the differences in the benchmark.

I updated the manuscript to consider your last comment in the following commit: https://github.com/VEZY/PlantSimEngine.jl/commit/0ff6b8f80ecc9a2efb22b033dd355f4a959dae09

This what I wrote:

>Cropbox.jl is a declarative crop modeling framework that share a common objective with PlantSimEngine: simplifying the process of defining models by eliminating the complexities of underlying simulations. While both frameworks can be applied across a wide range of scales, including land surface models (LSM), crop models, and functional-structural plant models (FSPM), they differ in their implementation strategies.
>
>Cropbox.jl defines a domain-specific language (DSL) tailored specifically for crop modeling, providing an intuitive approach for model definition. On the other hand, PlantSimEngine prioritizes speed as a fundamental aspect, ensuring efficient computations and enabling scalability right from the start. This is accomplished through meticulous considerations, such as type stability and pre-allocation to optimize performance. Type stability in PlantSimEngine ensures that variables maintain consistent types throughout computations, which leads to efficient execution. Additionally, the compatibility of PlantSimEngine with MultiScaleTreeGraph.jl [@vezy_multiscaletreegraphjl_2023] further enhances its scalability and facilitates multi-scale computations. MultiScaleTreeGraph.jl provides a powerful framework for representing and analyzing hierarchical structures, enabling PlantSimEngine to seamlessly handle complex systems spanning multiple scales.

Is this OK? I can change again if needed.
