### 2023-06-09 #5371 @crvernon
:wave: @VEZY could you fix the following please:

- [ ] Your Brisson et al. reference has an error in the doi link, see ... `https://doi.org/https://doi.org/10.1051/agro:19980501`.  Note the double `https//doi.org`.  You can remove that prefix in your bib file for the `doi` setting.
- [ ] You have `url` set for the Pradal et al. paper instead of `doi` which is causing our system to say that you are missing this DOI.  Could you change this in you bib file as a `doi` entry instead that only includes the `10.1071/FP08084` portion and not the full link?
- [ ] There is a DOI for the Carlson paper as well.  You can find it here:  https://arxiv.org/abs/2001.07625 .  Please add this too.

Otherwise the paper and package are looking really good!  Let me know when you have updated the above and we will move to the next step.  Thanks!
