### 2023-06-09 25812ce
Update paper.bib
Fix doi for several refs
