+++ docs/paper/paper.bib
@@ -24,7 +24,7 @@
 	volume = {18},
 	shorttitle = {{STICS}},
 	url = {https://hal.inrae.fr/hal-02698682},
-	doi = {https://doi.org/10.1051/agro:19980501},
+	doi = {10.1051/agro:19980501},
 	number = {5-6},
 	urldate = {2021-06-20},
 	journal = {Agronomie},
@@ -40,7 +40,7 @@ Publisher: EDP Sciences},
 @article{pradal_openalea_2008,
 	title = {{OpenAlea}: a visual programming and component-based software platform for plant modelling},
 	volume = {35},
-	url = {https://doi.org/10.1071/FP08084},
+	doi = {10.1071/FP08084},
 	abstract = {The development of functional–structural plant models requires an increasing amount of computer modelling. All these models are developed by different teams in various contexts and with different goals. Efficient and flexible computational frameworks are required to augment the interaction between these models, their reusability, and the possibility to compare them on identical datasets. In this paper, we present an open-source platform, OpenAlea, that provides a user-friendly environment for modellers, and advanced deployment methods. OpenAlea allows researchers to build models using a visual programming interface and provides a set of tools and models dedicated to plant modelling. Models and algorithms are embedded in OpenAlea ‘components’ with well defined input and output interfaces that can be easily interconnected to form more complex models and define more macroscopic components. The system architecture is based on the use of a general purpose, high-level, object-oriented script language, Python, widely used in other scientific areas. We present a brief rationale that underlies the architectural design of this system and we illustrate the use of the platform to assemble several heterogeneous model components and to rapidly prototype a complex modelling scenario.},
 	number = {10},
 	journal = {Functional Plant Biology},
@@ -134,6 +134,7 @@ Publisher: EDP Sciences},
 	title = {{MonteCarloMeasurements}.jl: {Nonlinear} {Propagation} of {Arbitrary} {Multivariate} {Distributions} by means of {Method} {Overloading}},
 	volume = {abs/2001.07625},
 	url = {https://arxiv.org/abs/2001.07625},
+	doi = {10.48550/arXiv.2001.07625},
 	journal = {CoRR},
 	author = {Carlson, Fredrik Bagge},
 	year = {2020},
