### 2023-06-09 #5371 @VEZY
@crvernon sorry for the mistakes on the doi, it should be OK now.

I have a question: as the FSPM community in Julia is taking some momentum, we agreed with another team to move our packages under a common organisation: https://github.com/VirtualPlantLab
I would like to put PlantSimEngine here, is this still possible to move the repo there and change the link in the paper ? I would understand if that's not possible, in this case I will keep the repo as is and provide a link to the new repo.
