+++ docs/paper/paper.bib
@@ -16,7 +16,7 @@
 	note = {Number: 11},
 	keywords = {Photosynthesis, Carbon dioxide, C3 photosynthesis, C4 photosynthesis, Curve fitting, Leaves, Stomata, Supply and demand},
 	pages = {e0143346},
-	file = {Duursma_2015_Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data.pdf:/Users/vezy/Zotero/storage/I7D8QT7U/Duursma_2015_Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data.pdf:application/pdf;Snapshot:/Users/vezy/Zotero/storage/EHXCTFFA/article.html:text/html},
+	file = {Duursma_2015_Plantecophys - An R Package for Analysing and Modelling Leaf Gas Exchange Data.},
 }
 
 @article{brisson_stics_1998,
@@ -34,7 +34,6 @@
 Publisher: EDP Sciences},
 	keywords = {BILAN AZOTE, RELATION PLANTE SOL},
 	pages = {311--346},
-	file = {Brisson et al_1998_STICS.pdf:/Users/vezy/Zotero/storage/DEZFQW8J/Brisson et al_1998_STICS.pdf:application/pdf},
 }
 
 @article{pradal_openalea_2008,
@@ -66,7 +65,6 @@ Publisher: EDP Sciences},
 	year = {2014},
 	keywords = {Biomass and nitrogen partitioning, Canopy dynamics, Integrated design environment, Phenological and morphological development, Reusable organ and function classes},
 	pages = {385--398},
-	file = {ScienceDirect Full Text PDF:/Users/vezy/Zotero/storage/KM94IK8P/Brown et al. - 2014 - Plant Modelling Framework Software for building a.pdf:application/pdf},
 }
 
 @article{holzworth_apsim_2014,
@@ -84,7 +82,6 @@ Publisher: EDP Sciences},
 	year = {2014},
 	keywords = {Simulation, Model, APSIM, Agricultural systems, Crop, Farming system, Gene-to-phenotype model},
 	pages = {327--350},
-	file = {ScienceDirect Snapshot:/Users/vezy/Zotero/storage/IHPDY4FF/S1364815214002102.html:text/html},
 }
 
 @article{jones_dssat_2003,
