### 2023-06-12 71f41dc
Update benchmark: parse the objects with a "$"
