### 2023-06-12 #5371 @VEZY
Hey @crvernon, I updated the paper because I found an error in the unit of one of the benchmark result. I re-ran the benchmark to be sure and updated the results accordingly. The values are slightly lower now because I forgot to [interpolate the global variables when doing the benchmark](https://juliaci.github.io/BenchmarkTools.jl/stable/manual/#Interpolating-values-into-benchmark-expressions).

Here's the commit I did: https://github.com/VirtualPlantLab/PlantSimEngine.jl/commit/71f41dcc4fb9bec5af877538ea72e407c5bf4e9d
