+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.8",
   "v0.6",
   "v0.4",
   "v0.3",
@@ -7,5 +8,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.6.0";
+var DOCUMENTER_NEWEST = "v0.8.1";
 var DOCUMENTER_STABLE = "stable";
