### 2023-06-15 85c2002
Update paper.md
Fix 2 from CO2
