### 2023-06-15 #5371 @VEZY
I just found a typo in CO2, the 2 is missing from the text. I did the correction in https://github.com/VirtualPlantLab/PlantSimEngine.jl/commit/85c20027aa5eab9ef1c883e2beebe41608fdbeaf
Do I have to make a new release on Github + Zenodo ?
