### 2023-06-26 #5371 @Kevin-Mattheus-Moerman
@VEZY I've inspected the paper, the software repository, and this review. Most is in order, however, below are some points that require your attention:

- [ ] Please edit the ZENODO archive listed license so that it matches your software license
- [ ] The paper currently lacks an acknowledgement section (which you may use for instance to acknowledge funding source). If this is in error you can add one now.
