### 2023-06-26 #5371 @VEZY
Hello @Kevin-Mattheus-Moerman, thanks for the information. I just updated the Zenodo repository to set the MIT licence.
There was no acknowledgement as I worked on this project alone, with no funding except for my salary paid by my institution. Several papers will follow with other people and funding though.

But now that you mention it, is it maybe customary to thank the reviewers and the editor of the paper?
